working on changes mentioned in PR
