+++ client/connect_loc.txt
@@ -1,1 +1,1 @@
-18.218.4.58
+localhost
